test(sidebar navigation): move tests to correct path, add assertion to check if ul is being rendered and change assertions testing child components functionality
