+++ app/components/navigation/__test__/FlowNavigation.test.tsx
@@ -3,7 +3,7 @@ import FlowNavigation from "~/components/navigation/FlowNavigation";
 import { NavState } from "~/services/navigation/navState";
 
 describe("FlowNavigation", () => {
-  it("renders a <nav> element with a <NavigationList> component", () => {
+  it("renders a navigation element", () => {
     const navItems = [
       {
         destination: "/destination",
@@ -16,8 +16,5 @@ describe("FlowNavigation", () => {
 
     const nav = screen.getByRole("navigation");
     expect(nav).toHaveClass("bg-white border-[1px] border-blue-400");
-
-    const list = screen.getByRole("list");
-    expect(list).toHaveClass("pl-0");
   });
 });
